+++ packages/remax-cli/src/__tests__/createHostComponent/toutiao.test.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import runTest from '../helpers/runTest';
-import { Platform } from '../../build/utils/platform';
+import { Platform } from '@remax/types';
 
 describe('create host components in toutiao', () => {
   runTest(