+++ packages/remax-cli/src/__tests__/fixtures/hook-config-babel/src/pages/index.js
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { View } from 'remax/ali';
+import styles from './index.less';
+
+function timesTwo(arr) {
+  return arr.map(n => n * 2);
+}
+
+export default () => {
+  console.log(timesTwo([1, 2, 3]));
+
+  return <View className={styles.text}>hello</View>;
+};