+++ packages/remax-cli/src/__tests__/nativeComponent/ali.test.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import runTest from '../helpers/runTest';
-import { Platform } from '../../build/utils/platform';
+import { Platform } from '@remax/types';
 
 describe('use native components in ali app', () => {
   const cwd = path.resolve(__dirname, '../fixtures/nativeComponent/expected/ali');