+++ packages/remax-cli/src/__tests__/plugin.test.ts
@@ -1,0 +1,7 @@
+import runTest from './helpers/runTest';
+
+describe('plugin hooks', () => {
+  runTest('hook-config-webpack');
+
+  runTest('hook-config-babel');
+});