+++ packages/remax-cli/src/__tests__/wechat.test.ts
@@ -1,5 +1,5 @@
 import runTest from './helpers/runTest';
-import { Platform } from '../build/utils/platform';
+import { Platform } from '@remax/types';
 
 describe('build wechat app', () => {
   runTest('wechat', Platform.wechat);