+++ packages/remax-cli/src/getConfig.ts
@@ -1,41 +1,20 @@
 import fs from 'fs';
 import path from 'path';
-import esm from 'esm';
 import { getDefaultOptions } from './defaultOptions';
-import { RemaxOptions, RemaxConfig } from '@remax/types';
+import { Options } from '@remax/types';
 import validateOptions from 'schema-utils';
-import schema from './RemaxOptionsSchema.json';
-import API from './API';
+import schema from './OptionsSchema.json';
 
 export interface CliOptions {
   target: string;
 }
 
 function readJavascriptConfig(path: string) {
-  // eslint-disable-next-line
-  require = esm(module, {
-    cjs: {
-      dedefault: true,
-    },
-  });
-  delete require.cache[require.resolve(path)];
   const config = require(path);
-
   return config || {};
 }
 
-function validateTurboPages(config: RemaxConfig) {
-  // adapter name 还不存在，说明 adapter plugin 还没有初始化
-  if (!API.adapter.target || API.adapter.target === 'ali') {
-    return;
-  }
-
-  if (config.turboPages && config.turboPages.length > 0) {
-    throw new Error('turboPages 目前仅支持 ali 平台开启');
-  }
-}
-
-export default function getConfig(validate = true): RemaxOptions {
+export default function getConfig(validate = true): Options {
   const configPath: string = path.join(process.cwd(), './remax.config');
 
   let options = {};
@@ -48,8 +27,6 @@ export default function getConfig(validate = true): RemaxOptions {
     validateOptions(schema as any, options, {
       name: 'remax',
     });
-
-    validateTurboPages(options);
   }
 
   return {