+++ packages/remax-cli/src/index.ts
@@ -1,4 +1,3 @@
-import API from './API';
 export { default as logger } from './build/utils/output';
 import yargs from 'yargs';
 import { build } from './build';
@@ -12,7 +11,7 @@ export { getDefaultOptions } from './defaultOptions';
 
 export function run(args: any, callback?: yargs.ParseCallback) {
   cli = yargs
-    .scriptName('remax-cli')
+    .scriptName('remax')
     .usage('Usage: $0 <command> [options]')
     .command<any>(
       'build',
@@ -26,8 +25,6 @@ export function run(args: any, callback?: yargs.ParseCallback) {
         build(argv);
       }
     );
-
-  cli = API.extendsCLI({ cli });
 
   cli
     .option('watch', {